move modern AssetListItem
