+++ src/components/lists/AssetListItem/AssetListItem.js
